UseGraphSonXX-methods must be on the environment. Anything else is not obvious.
