+++ test/ExRam.Gremlinq.PublicApi.Tests/PublicApiTests.SupportNewtonsoftJson.verified.cs
@@ -10,13 +10,7 @@ namespace ExRam.Gremlinq.Core
     public static class GremlinQueryEnvironmentExtensions
     {
         public static ExRam.Gremlinq.Core.IGremlinQueryEnvironment StoreTimeSpansAsNumbers(this ExRam.Gremlinq.Core.IGremlinQueryEnvironment environment) { }
-    }
-}
-namespace ExRam.Gremlinq.Core.Transformation
-{
-    public static class TransformerExtensions
-    {
-        public static ExRam.Gremlinq.Core.Transformation.ITransformer UseGraphSon2(this ExRam.Gremlinq.Core.Transformation.ITransformer transformer) { }
-        public static ExRam.Gremlinq.Core.Transformation.ITransformer UseGraphSon3(this ExRam.Gremlinq.Core.Transformation.ITransformer transformer) { }
+        public static ExRam.Gremlinq.Core.IGremlinQueryEnvironment UseGraphSon2(this ExRam.Gremlinq.Core.IGremlinQueryEnvironment environment) { }
+        public static ExRam.Gremlinq.Core.IGremlinQueryEnvironment UseGraphSon3(this ExRam.Gremlinq.Core.IGremlinQueryEnvironment environment) { }
     }
 }
